requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.3
python-dotenv==1.0.0
black==23.7.0